+++ RELEASES.md
@@ -1,14 +1,25 @@
 ## Release Notes
 
 
-### hunger (v0.8.0) -- in progress
+### tick (v0.9.0) -- in progress
+
+- the World shall tick, and all things shall proceed from there
+- the main() loop shall be simplified further
+- Food shall have tick
+- World shall have tick
+- Groc shall have observe/decide/act
+- Genetic metabolism factors - randomized
+- Genetic eating factors - randomized
+- Variable food value - randomized
+
+### hunger (v0.8.0) -- complete
  
 - HUNGRY as a state of being
 - Food as a class of item in the World
-- Food Points and Hit Points
-- Genetic metabolism factors
-- Genetic eating factors
-- Variable food value (calories per bite)
+- Food Points and Calories
+- Genetic metabolism factors (constant for now but easily randomizable)
+- Genetic eating factors (constant for now but easily randomizable)
+- Variable food value (constant for now but easily randomizable)
 
 ### uno (v0.7.0) -- complete